% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson-Multinomial_functions.R
\name{dpmd}
\alias{dpmd}
\title{Probability Mass Function of Poisson-Multinomial Distributions}
\usage{
dpmd(pmat, x = c(0, 0, 0, 0), method = "DFT-CF", B = 10^3)
}
\arguments{
\item{pmat}{A matrix of probabilities. Each row of pmat should add up 
to 1.}

\item{x}{Result vector(probability mass point) specified by user when 
the selected method is "SIM" or "NA". The vector 
\code{x = (x_{1}, x_{2}, \ldots)} is used for computing 
\eqn{P(X_{1}=x_{1}, X_{2}=x_{2}, \ldots)}.}

\item{method}{Character string stands for the method selected by user to 
compute the probability mass. The method can only be one of 
the following four: 
\code{"DFT-CF"},
\code{"NA"},
\code{"SIM"},
\code{"SIM-ALL"}.}

\item{B}{Simulation repeating time. Will be ignored if users do not
choose \code{"SIM-ALL"} or \code{"SIM"}
as method.}
}
\value{
For a single probability mass point, \code{dpmd} returns a probability value. 

For all probability mass points of a given \code{pmat}, it returns a 
multi-dimensional array. For instance, for the \code{pmat} matrix in the 
following example, the value of the array element \eqn{a_{1,2,1}} = 0.90 means 
the value of probability mass point (0,1,0,2) is 0.90.
}
\description{
Probability mass function of Poisson-Multinomial distributions 
specified by input matrix and computed through selected method. This function 
is capable for computation of the whole probability mass function as well as
of one single probability mass point.
}
\details{
For the methods we applied in \code{dpmd}, \code{"DFT-CF"} is an exact method 
to calculate all probability mass points of Poisson-Multinomial Distributions
via FFT algorithm. When users select \code{"DFT-CF"}, \code{dpmd} will ignore
\code{vec} and output the whole probability mass function.

\code{"SIM-ALL"} is a simulation method using naive simulation scheme to 
calculate the whole probability mass function, under this selection the input 
of \code{vec} will be ignore. Notice the accuracy and running time will be
effected by user choice of \code{B}. Usually \code{B}=10^5 or 10^6 will be 
accurate enough. Increasing \code{B} to larger than 10^8 will heavily aggravate 
computation burden of a CPU or GPU. 

Given \code{pmat} with dimension \eqn{n \times m}, the number of total probability 
mass points is \eqn{(n+1)^{m-1}}. Thus when the dimension of \code{pmat} 
increases and the users selected method is one of \code{"DFT-CF"} and 
\code{"SIM-ALL"}, the computation burden of \code{dpmd} might challenge the 
capability of a computer because both of the methods calculate all probability 
mass points of Poisson-Multinomial distributions.

 
\code{"SIM"} is as same as \code{"SIM-ALL"} except that it only computes a 
single probability mass point specified by \code{vec}.

\code{"NA"} specifies an approximation method using Normal approximation to 
compute the probability mass point of the \code{vec} vector
input by user.
}
\examples{
pp=matrix(c(.1, .1, .1, .7, .1, .3, .3, .3, .5, .2, .1, .2), nrow = 3, byrow = TRUE)


dpmd(pmat = pp)
dpmd(pmat = pp, method = "SIM-ALL", B = 10^3)
dpmd(pmat = pp, x = c(0,0,1,2), method = "NA" )
dpmd(pmat = pp, x = c(0,0,1,2), method = "SIM", B = 10^3)

}
