% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dpmn}
\alias{dpmn}
\title{probability density of Poisson multinomial distribution}
\usage{
dpmn(pp, method = "DFT-CF", vec = c(0, 0, 0, 0, 0), t = 100)
}
\arguments{
\item{pp}{The pp is a probability matrix which will be input by user}

\item{method}{method selected by user to compute the probability mass}

\item{vec}{result vec input by user}

\item{t}{simulation repeat time}
}
\value{
The probability mass of PMD
}
\description{
probability density of Poisson multinomial distribution
}
\examples{
aa=matrix(c(.1, .1, .1, .7, .1, .3, .3, .3, .5, .2, .1, .2, .5, .1, .1, .3), nrow=4, byrow=TRUE)
pp=aa[1:3,]
dpmn(pp)
}
