% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmd.R
\name{getchain}
\alias{getchain}
\title{Get reaction chain for specific mass to charge ratio}
\usage{
getchain(list, diff, mass, accuracy = 4, ...)
}
\arguments{
\item{list}{a list with mzrt profile}

\item{diff}{paired mass distance(s) of interests}

\item{mass}{a specific mass for known compound}

\item{accuracy}{measured mass or mass to charge ratio in digits, default 4}

\item{...}{other parameters for getpmd}
}
\value{
a list with mzrt profile and reaction chain dataframe
}
\description{
Get reaction chain for specific mass to charge ratio
}
\examples{
data(spmeinvivo)
# check metabolites of C18H39NO
pmd <- getchain(spmeinvivo,diff = c(2.02,14.02,15.99,58.04,13.98),mass = 286.3101)
}
