% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmd.R
\name{getcluster}
\alias{getcluster}
\title{Get Pseudo-Spectrum as peaks cluster based on pmd analysis.}
\usage{
getcluster(list, corcutoff = NULL)
}
\arguments{
\item{list}{a list from getstd function}

\item{corcutoff}{cutoff of the correlation coefficient, default NULL}
}
\value{
list with Pseudo-Spectrum index
}
\description{
Get Pseudo-Spectrum as peaks cluster based on pmd analysis.
}
\examples{
\donttest{
data(spmeinvivo)
re <- getstd(spmeinvivo)
cluster <- getcluster(re)
}
}
