% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmd.R
\name{getcorcluster}
\alias{getcorcluster}
\title{Get Pseudo-Spectrum as peaks cluster based on correlation analysis.}
\usage{
getcorcluster(list, corcutoff = 0.9, rtcutoff = 10)
}
\arguments{
\item{list}{a list with peaks intensity}

\item{corcutoff}{cutoff of the correlation coefficient, default 0.9}

\item{rtcutoff}{cutoff of the distances in cluster, default 10}
}
\value{
list with Pseudo-Spectrum index
}
\description{
Get Pseudo-Spectrum as peaks cluster based on correlation analysis.
}
\examples{
\donttest{
data(spmeinvivo)
cluster <- getcorcluster(spmeinvivo)
}
}
