% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalstd.R
\name{getcorpseudospectrum}
\alias{getcorpseudospectrum}
\title{Get Pseudo-Spectrum as peaks cluster based on correlation analysis.}
\usage{
getcorpseudospectrum(list, corcutoff = 0.9, rtcutoff = 10, accuracy = 4)
}
\arguments{
\item{list}{a list with peaks intensity}

\item{corcutoff}{cutoff of the correlation coefficient, default 0.9}

\item{rtcutoff}{cutoff of the distances in cluster, default 10}

\item{accuracy}{measured mass or mass to charge ratio in digits, default 4}
}
\value{
list with Pseudo-Spectrum index
}
\description{
Get Pseudo-Spectrum as peaks cluster based on correlation analysis.
}
\examples{
data(spmeinvivo)
pseudo <- getcorpseudospectrum(spmeinvivo)
}
