% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmdanno.R
\name{getms2pmd}
\alias{getms2pmd}
\title{read in MSP file as list for ms/ms annotation}
\usage{
getms2pmd(file, digits = 2, icf = 10)
}
\arguments{
\item{file}{the path to your MSP file}

\item{digits}{mass or mass to charge ratio accuracy for pmd, default 2}

\item{icf}{intensity cutoff, default 10 percentage}
}
\value{
list a list with MSP information for MS/MS annotation
}
\description{
read in MSP file as list for ms/ms annotation
}
