% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmdanno.R
\name{getmspmd}
\alias{getmspmd}
\title{read in MSP file as list for EI-MS annotation}
\usage{
getmspmd(file, digits = 2)
}
\arguments{
\item{file}{the path to your MSP file}

\item{digits}{mass or mass to charge ratio accuracy for pmd, default 0}

\item{icf}{intensity cutoff, default 10 percentage}
}
\value{
list a list with MSP information for EI-MS annotation
}
\description{
read in MSP file as list for EI-MS annotation
}
