% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalstd.R
\name{getnse}
\alias{getnse}
\title{Calculate Normalized Spectral Entropy (NSE)}
\usage{
getnse(x)
}
\arguments{
\item{x}{Matrix or dataframe with samples in column and features in row}
}
\value{
Normalized Spectral Entropy (NSE) between 0 and 1
}
\description{
Calculate Normalized Spectral Entropy (NSE)
}
\examples{
data(spmeinvivo)
getnse(spmeinvivo$data)
}
