% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmd.R
\name{getpmd}
\alias{getpmd}
\title{Get pmd for specific reactions}
\usage{
getpmd(list, pmd, rtcutoff = 10)
}
\arguments{
\item{list}{a list with mzrt profile}

\item{pmd}{a specific paired mass distances}

\item{rtcutoff}{cutoff of the distances in retention time hierarchical clustering analysis, default 10

#' @return list with tentative isotope, adducts, and neutral loss peaks' index, retention time clusters.}
}
\description{
Get pmd for specific reactions
}
\examples{
\donttest{
data(spmeinvivo)
pmd <- getpmd(spmeinvivo,pmd=15.99)
}
}
\seealso{
\code{\link{getpaired}},\code{\link{getstd}},\code{\link{getsda}},\code{\link{getrda}}
}
