% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmd.R
\name{getposneg}
\alias{getposneg}
\title{Link pos mode peak list with neg mode peak list by pmd.}
\usage{
getposneg(pos, neg, pmd = 2.02, digits = 2)
}
\arguments{
\item{pos}{a list with mzrt profile collected from positive mode.}

\item{neg}{a list with mzrt profile collected from negative mode.}

\item{pmd}{numeric or numeric vector}

\item{digits}{mass or mass to charge ratio accuracy for pmd, default 2}
}
\value{
dataframe with filtered postive and negative peak list
}
\description{
Link pos mode peak list with neg mode peak list by pmd.
}
