% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalstd.R
\name{getpseudospectrum}
\alias{getpseudospectrum}
\title{Get Pseudo-Spectrum as peaks cluster based on correlation analysis.}
\usage{
getpseudospectrum(
  list,
  corcutoff = NULL,
  rtcutoff = 10,
  accuracy = 4,
  ng = NULL,
  digits = 2
)
}
\arguments{
\item{list}{A list containing peak intensities, m/z values, and retention times. Must include elements: mz, rt, data.}

\item{corcutoff}{Cutoff value for correlation coefficient (default: 0.9).}

\item{rtcutoff}{Cutoff value for retention time clustering (default: 10).}

\item{accuracy}{Number of decimal places for m/z rounding (default: 4).}

\item{ng}{cutoff of global PMD's retention time group numbers, If ng = NULL, 20 percent of RT cluster will be used as ng, default NULL.}

\item{digits}{mass or mass to charge ratio accuracy for pmd, default 2}
}
\value{
A list with pseudo-spectrum clustering results.
}
\description{
Get Pseudo-Spectrum as peaks cluster based on correlation analysis.
}
\examples{
data(spmeinvivo)
pseudo <- getpseudospectrum(spmeinvivo)
}
