% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalstd.R
\name{getsim}
\alias{getsim}
\title{Evaluate the similarity of two matrices or dataframes}
\usage{
getsim(x, y, dim = NULL)
}
\arguments{
\item{x}{Matrix or dataframe with samples in column and features in row (Original data)}

\item{y}{Matrix or dataframe with samples in column and features in row (Reduced data)}

\item{dim}{number of retained dimensions for PCASF. Defaults to all.}
}
\value{
A numeric vector containing PCASF, RV coefficient, and Mantel test r (Pearson and Spearman).
}
\description{
Evaluate the similarity of two matrices or dataframes
}
\examples{
data(spmeinvivo)
re <- globalstd(spmeinvivo)
x <- spmeinvivo$data
y <- spmeinvivo$data[re$stdmassindex, ]
getsim(x, y)
}
