% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmdb}
\alias{hmdb}
\title{A dataframe containing HMDB with unique accurate mass pmd with three digits frequency larger than 1 and accuracy percentage larger than 0.9.}
\format{
A dataframe with atoms numbers of C, H, O, N, P, S
\describe{
  \item{percentage}{accuracy of atom numbers prediction}
  \item{pmd2}{pmd with two digits}
  \item{pmd}{pmd with three digits}
  }
}
\usage{
data(hmdb)
}
\description{
A dataframe containing HMDB with unique accurate mass pmd with three digits frequency larger than 1 and accuracy percentage larger than 0.9.
}
\keyword{datasets}
