% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmdbp}
\alias{hmdbp}
\title{A list dataset containing HMDB unique accurate mass pmd analysis results}
\format{A list with two vectors
\describe{
  \item{massp}{all unique hmdb mass probability across all pmds}
  \item{pmdp}{pmds probability across all unique hmdb mass}
  }}
\usage{
data(hmdbp)
}
\description{
A list dataset containing HMDB unique accurate mass pmd analysis results
}
\keyword{datasets}
