% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{keggrall}
\alias{keggrall}
\title{A dataframe containing reaction related accurate mass pmd and related reaction formula with KEGG ID}
\format{A dataframe with KEGG reaction, their realted pmd and atoms numbers of C, H, O, N, P, S
\describe{
  \item{ID}{KEGG reaction ID}
  \item{pmd}{pmd with three digits}
  }}
\usage{
data(keggrall)
}
\description{
A dataframe containing reaction related accurate mass pmd and related reaction formula with KEGG ID
}
\keyword{datasets}
