% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mu.cmpt}
\alias{mu.cmpt}
\title{mu.cmpt}
\usage{
mu.cmpt(p)
}
\arguments{
\item{p}{The p is a probability vmatrix which will be input by user}
}
\value{
The probability mass of mean mu
}
\description{
mu.cmpt
}
