% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omics}
\alias{omics}
\title{A dataframe containing multiple reaction database ID and their related accurate mass pmd and related reactions}
\format{A dataframe with reaction and their realted pmd
\describe{
  \item{KEGG}{KEGG reaction ID}
  \item{RHEA_ID}{RHEA_ID}
  \item{DIRECTION}{reaction direction}
  \item{MASTER_ID}{master reaction RHEA ID}
  \item{ec}{ec reaction ID}
  \item{ecocyc}{ecocyc reaction ID}
  \item{macie}{macie reaction ID}
  \item{metacyc}{metacyc reaction ID}
  \item{reactome}{reactome reaction ID}
  \item{compounds}{reaction related compounds}
  \item{pmd}{pmd with two digits}
  \item{pmd2}{pmd with three digits}
  }}
\usage{
data(omics)
}
\description{
A dataframe containing multiple reaction database ID and their related accurate mass pmd and related reactions
}
\keyword{datasets}
