% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmdvis.R
\name{plotcn}
\alias{plotcn}
\title{plot PMD KEGG network for certain compounds and output network average distance and degree}
\usage{
plotcn(formula, name, pmd)
}
\arguments{
\item{formula}{Chemical formula}

\item{name}{Compound name}

\item{pmd}{specific paired mass distances}
}
\description{
plot PMD KEGG network for certain compounds and output network average distance and degree
}
\examples{
plotcn('C6H12O6','Glucose',c(2.016,14.016,15.995))
}
