% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmdanno.R
\name{plotpmdanno}
\alias{plotpmdanno}
\title{Show MS/MS pmd annotation result from `pmdanno` function}
\usage{
plotpmdanno(anno)
}
\arguments{
\item{anno}{list from pmdanno function}
}
\description{
Show MS/MS pmd annotation result from `pmdanno` function
}
