% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmd.R
\name{plotstdrt}
\alias{plotstdrt}
\title{Plot the std mass from GlobalStd algorithm in certain retention time groups}
\usage{
plotstdrt(list, rtcluster, ...)
}
\arguments{
\item{list}{a list from getstd function}

\item{rtcluster}{retention time group index}

\item{...}{other parameters for plot function}
}
\description{
Plot the std mass from GlobalStd algorithm in certain retention time groups
}
\examples{
\donttest{
data(spmeinvivo)
pmd <- getpaired(spmeinvivo)
std <- getstd(pmd)
plotstdrt(std,rtcluster = 6)
}
}
\seealso{
\code{\link{getstd}}, \code{\link{globalstd}},\code{\link{plotstd}},\code{\link{plotpaired}},\code{\link{plotstdsda}}
}
