% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmdvis.R
\name{plotstdsda}
\alias{plotstdsda}
\title{Plot the std mass from GlobalStd algorithm in structure directed analysis(SDA) groups}
\usage{
plotstdsda(list, index = NULL, ...)
}
\arguments{
\item{list}{a list from getsda function}

\item{index}{index for PMD value}

\item{...}{other parameters for plot function}
}
\description{
Plot the std mass from GlobalStd algorithm in structure directed analysis(SDA) groups
}
\examples{
data(spmeinvivo)
re <- globalstd(spmeinvivo, sda=TRUE)
plotstdsda(re)
}
\seealso{
\code{\link{getstd}}, \code{\link{globalstd}},\code{\link{plotstd}},\code{\link{plotpaired}},\code{\link{plotstdrt}}
}
