% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pm_simulation_arma}
\alias{pm_simulation_arma}
\title{pm_simulation}
\usage{
pm_simulation_arma(pp, x_vec, t)
}
\arguments{
\item{pp}{probability matrix}

\item{x_vec}{result vector}

\item{t}{simulation time}
}
\value{
res probability of x_vec
}
\description{
pm_simulation
}
