% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson-Multinomial_functions.R
\name{pmatrix}
\alias{pmatrix}
\title{pmatrix}
\usage{
pmatrix(n, m)
}
\arguments{
\item{n}{column dimension}

\item{m}{row dimension}
}
\value{
a randomly generated Poisson multinomial distribution probability matrix
}
\description{
pmatrix
}
\examples{
pp = pmatrix(2,2)
pp
}
