% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pmd.by.demands}
\alias{pmd.by.demands}
\title{PMN density calculated by input vector}
\usage{
pmd.by.demands(x_vec, pp, t = 1000)
}
\arguments{
\item{x_vec}{result vec input by user}

\item{pp}{The pp is a probability matrix which will be input by user}

\item{t}{simulation repeat time}
}
\value{
The probability mass of PMD
}
\description{
PMN density calculated by input vector
}
\examples{
aa=matrix(c(.1, .1, .1, .7, .1, .3, .3, .3, .5, .2, .1, .2, .5, .1, .1, .3), nrow=4, byrow=TRUE)
pp=aa[1:3,]
pmd.by.demands(c(1,2,0,0),pp,t=10^5)
}
