% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{pmd}
\alias{pmd}
\title{main function}
\usage{
pmd(pp, method = "DFT-CF", vec = c(0, 0, 0, 0, 0), t = 100)
}
\arguments{
\item{pp}{a matrix of input}
}
\value{
res a result array
}
\description{
main function
}
