% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmdanno.R
\name{pmdanno}
\alias{pmdanno}
\title{Perform MS/MS pmd annotation for mgf file}
\usage{
pmdanno(file, db = "qtof", ppm = 10, prems = 1.1, pmdc = 0.6)
}
\arguments{
\item{file}{mgf file generated from MS/MS data}

\item{db}{database could be 'qtof', 'qqq', 'orb' data base from hmdb or list object from `getms2pmd`}

\item{ppm}{mass accuracy, default 10}

\item{prems}{precersor mass range, default 1.1 to include M+H or M-H}

\item{pmdc}{pmd length percentage cutoff for annotation. 0.6(default) means 60 percentage of the pmds in your sample could be found in certain compound pmd database}
}
\value{
list with HMDB annotation results
}
\description{
Perform MS/MS pmd annotation for mgf file
}
\examples{
file <- system.file("extdata", "challenge-msms.mgf", package = "pmd")
anno <- pmdanno(file)
}
