% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pmn_simulation_arma}
\alias{pmn_simulation_arma}
\title{pmn_simulation}
\usage{
pmn_simulation_arma(pp, nnt, l_vec, cn_vec, t)
}
\arguments{
\item{pp}{probability matrix}

\item{nnt}{number of outcomes}

\item{l_vec}{result vector}

\item{cn_vec}{cn vector}

\item{t}{simulation time}
}
\value{
res array of probabilities
}
\description{
pmn_simulation
}
