% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson-Multinomial_functions.R
\name{PoissonMultinomial-distribution}
\alias{PoissonMultinomial-distribution}
\alias{dpmd}
\alias{ppmd}
\alias{rpmd}
\title{Probability Mass Function of Poisson-Multinomial Distributions}
\usage{
dpmd(pp, method = "DFT-CF", vec = c(0, 0, 0, 0, 0), B = 100)

ppmd(pp, x, method = "DFT-CF", B = 1000)

rpmd(pp, n)
}
\arguments{
\item{pp}{A matrix of probabilities. Each row of pp should add up 
to 1.}

\item{method}{Character string stands for the method selected by user to 
compute the probability mass. There are totally 4 methods.

\code{"DFT-CF"}: An exact method to calculate all 
probability mass points of Poisson-Multinomial Distributions
via FFT algorithm.

\code{"simulation"}: A simulation method calculating all 
probability mass points.

\code{"NA by demands"}: An approximation 
method using Normal approximation to compute the 
probability for the 'vec' vector input by user.

\code{"simulation by demands"}: The same simulation method 
as above just to compute single probability mass point as 
input by user.}

\item{vec}{Result vector(probability mass point) specified by user.
Eg. pp is 4 by 3 matrix then a user might be interested in 
the probability of getting result: vec=c(0,0,1,2).}

\item{B}{Simulation repeating time.}

\item{x}{Vector \eqn{x = (x_{1},x_{2},\ldots)} for computing 
\eqn{P(X_{1} \leq x_{1},X_{2} \leq x_{2},\ldots)}.}

\item{n}{Number of samples to be generated.}
}
\value{
For a single mass point, \code{dpmd} returns a probability. 


For all probability mass points of a given \code{pp}, it returns a 
multi-dimensional array. For instance, for the \code{pp} matrix in the 
following example, the value of the
array element \eqn{a_{1,2,1}} = 0.90 means the value of probability mass point 
(0,1,0,2) is 0.90. Notice method \code{"DFT-CF"} can not return a probability 
of single mass point.
                   
                   
Given matrix \code{pp}, vector \code{x} and method, \code{ppmd} returns 
\eqn{P(X_{1} \leq x_{1},X_{2} \leq x_{2},\ldots)}.
                   
                   
\code{rpmd} gives a matrix of generated samples with given \code{pp}. 
Each row represents a sample.
}
\description{
Probability mass function, 
cumulative distribution function, 
and a random number generator of Poisson-Multinomial distributions.
}
\examples{
pp=matrix(c(.1, .1, .1, .7, .1, .3, .3, .3, .5, .2, .1, .2), nrow=3, byrow=TRUE)


dpmd(pp)
dpmd(pp,"simulation",B=10^3)
dpmd(pp,"NA by demands", vec = c(0,0,1,2))
dpmd(pp,"simulation by demands", vec = c(0,0,1,2), B=10^3)


ppmd(pp, x = c(3,2,1,3))
ppmd(pp, x = c(3,2,1,3), method = "simulation", B = 10^3)
ppmd(pp, x = c(3,2,1,3), method = "NA")



rpmd(pp,5)
}
