% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson-Multinomial_functions.R
\name{ppmd}
\alias{ppmd}
\title{cumulative distribution function of PMN}
\usage{
ppmd(pp, x, method = "DFT-CF", B = 1000)
}
\arguments{
\item{pp}{input matrix of probabilities}

\item{x}{input vector}

\item{method}{method}

\item{B}{repeating time}
}
\value{
prob
}
\description{
cumulative distribution function of PMN
}
\examples{
aa=matrix(c(.1, .1, .1, .7, .1, .3, .3, .3, .5, .2, .1, .2, .5, .1, .1, .3), nrow=4, byrow=TRUE)
pp=aa[1:3,]
ppmd(pp,c(3,2,1,3))
}
