% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ppmn}
\alias{ppmn}
\title{cumulative distribution function of PMN}
\usage{
ppmn(pp, x, method = "DFT-CF", t = 1000)
}
\arguments{
\item{pp}{input matrix of probabilities}

\item{x}{input vector}

\item{method}{method}

\item{t}{repeating time}
}
\value{
prob
}
\description{
cumulative distribution function of PMN
}
\examples{
aa=matrix(c(.1, .1, .1, .7, .1, .3, .3, .3, .5, .2, .1, .2, .5, .1, .1, .3), nrow=4, byrow=TRUE)
pp=aa[1:3,]
ppmn(pp,c(3,2,1,3))
}
