% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qtof}
\alias{qtof}
\title{A list containing HMDB qtof MS/MS data with peaks larger than 10 percentage for PMD annatation}
\format{A list containing HMDB qtof MS/MS data with peaks larger than 10 percentage for PMD annatation
\describe{
  \item{name}{HMDB ID}
  \item{mz}{mass to charge ratio}
  \item{msms}{msms pmd}
  \item{msmsraw}{raw msms data}
  }}
\usage{
data(qtof)
}
\description{
A list containing HMDB qtof MS/MS data with peaks larger than 10 percentage for PMD annatation
}
\keyword{datasets}
