% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmultinom_1}
\alias{rmultinom_1}
\title{rmultinom_1 method}
\usage{
rmultinom_1(size, probs, N)
}
\arguments{
\item{size}{vector of sizes}

\item{probs}{probability vector}

\item{N}{number of samples}
}
\value{
outcome
}
\description{
rmultinom_1 method
}
