% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmultinom_rcpp}
\alias{rmultinom_rcpp}
\title{rmultinom}
\usage{
rmultinom_rcpp(n, size, probs)
}
\arguments{
\item{n}{length of probability vector}

\item{size}{vector of sizes}

\item{probs}{probabilities}
}
\value{
sim simulation matrix
}
\description{
rmultinom
}
