% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson-Multinomial_functions.R
\name{rpmd}
\alias{rpmd}
\title{Poisson-Multinomial Distribution Random Number Generator}
\usage{
rpmd(pmat, n = 1)
}
\arguments{
\item{pmat}{A \eqn{n \times m} matrix of probabilities. \eqn{n} is the number of independent trials.
\eqn{m} is the number of categories.
Each row of pmat describes the success probability for the corresponding
trial and it should add up to 1.}

\item{n}{Number of samples to be generated.}
}
\value{
A \eqn{n \times m} matrix of samples, each row stands for one draw from PMD with success probability matrix \code{pmat}.
}
\description{
Generating random samples of given a Poisson-Multinomial distribution.
}
\examples{
pp=matrix(c(.1, .1, .1, .7, .1, .3, .3, .3, .5, .2, .1, .2), nrow = 3, byrow = TRUE)


rpmd(pmat = pp, n = 5)

}
