% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rpmd_arma}
\alias{rpmd_arma}
\title{rpmn}
\usage{
rpmd_arma(pp)
}
\arguments{
\item{pp}{probability matrix}
}
\value{
finalres a sample of Poisson-Multinomial distribution given probability matrix pp.
}
\description{
rpmn
}
