% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmd.R
\docType{data}
\name{sda}
\alias{sda}
\title{A dataset containing common Paired mass differeces of substructure, ions replacements, and reaction}
\format{A data frame with 94 rows and 4 variables:
\describe{
  \item{PMD}{Paired mass differeces}
  \item{origin}{potentical sources}
  \item{Ref.}{references}
  \item{mode}{positive, negative or both mode to find corresponding PMDs}
  }}
\usage{
data(sda)
}
\description{
A dataset containing common Paired mass differeces of substructure, ions replacements, and reaction
}
\keyword{datasets}
