% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sig.cmpt}
\alias{sig.cmpt}
\title{sig.cmpt}
\usage{
sig.cmpt(p)
}
\arguments{
\item{p}{The p is a probability vmatrix which will be input by user}
}
\value{
The probability mass of variance sigma
}
\description{
sig.cmpt
}
