% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spmeinvivo}
\alias{spmeinvivo}
\title{A peaks list dataset containing 9 samples from 3 fish with triplicates samples for each fish from LC-MS.}
\format{
A list with 4 variables from 1459 LC-MS peaks:
\describe{
  \item{mz}{mass to charge ratios}
  \item{rt}{retention time}
  \item{data}{intensity matrix}
  \item{group}{group information}
  }
}
\usage{
data(spmeinvivo)
}
\description{
A peaks list dataset containing 9 samples from 3 fish with triplicates samples for each fish from LC-MS.
}
\keyword{datasets}
