% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson-Multinomial_functions.R
\name{dpmd}
\alias{dpmd}
\title{Probability Mass Function of Poisson-Multinomial Distributions}
\usage{
dpmd(pmat, x = c(0, 0, 0, 0), method = "DFT-CF", B = 10^3)
}
\arguments{
\item{pmat}{A \eqn{n \times m} matrix of probabilities. \eqn{n} is the number of independent trials.
\eqn{m} is the number of categories. Also called success probability matrix.
Each row of pmat describes the success probability for the corresponding
trial and it should add up to 1.}

\item{x}{Result vector of length \eqn{m} (probability mass point) specified by user when 
the selected method is "SIM" or "NA". The vector 
\eqn{x = (x_{1}, x_{2}, \ldots, x_{m})} is used for computing 
\eqn{P(X_{1}=x_{1}, X_{2}=x_{2}, \ldots, X_{m} = x_{m})}.}

\item{method}{Character string stands for the method selected by user to 
compute the probability mass. The method can only be one of 
the following four: 
\code{"DFT-CF"},
\code{"NA"},
\code{"SIM"},
\code{"SIM-ALL"}.}

\item{B}{Number of repetitions in the simulation method. Will be ignored if users do not
choose \code{"SIM-ALL"} or \code{"SIM"} method.}
}
\value{
For a single mass point, \code{dpmd} returns the probability mass function at that point. 

For all mass points of a given \code{pmat}, it returns a 
multi-dimensional array. For instance, for the \code{pmat} matrix in the 
following example, the value of the array element \eqn{a_{1,2,1}} = 0.90 means 
the value of probability mass point (0,1,0,2) is 0.90.
}
\description{
Probability mass function of Poisson-Multinomial distributions 
specified by input matrix and computed through selected method. This function 
is capable for computation of the whole probability mass function as well as
of one single probability mass point.
}
\details{
Given \code{pmat} with dimension \eqn{n \times m}, the total number of outcomes is \eqn{(n+1)^{m-1}}. 
For the methods we applied in \code{dpmd}, \code{"DFT-CF"} is an exact method 
to calculate all mass points of Poisson-Multinomial Distributions
via FFT algorithm. When users select \code{"DFT-CF"}, \code{dpmd} will ignore
\code{vec} and return the probability mass function for all outcomes.

\code{"SIM-ALL"} is a simulation method using a naive simulation scheme to 
calculate the whole probability mass function. Under this selection, the input 
of \code{vec} will be ignored. Notice that the accuracy and running time will be
affected by user choice of \code{B}. Usually \code{B}=10^5 or 10^6 will be 
accurate enough. Increasing \code{B} to larger than 10^8 will heavily aggravate 
computational burden of a CPU or GPU. 

When the dimension of \code{pmat} increases, the computation burden of \code{"DFT-CF"} and 
\code{"SIM-ALL"} method might challenge the 
capability of a computer because both of the methods calculate all
mass points of Poisson-Multinomial distributions.

 
\code{"SIM"} is as same as \code{"SIM-ALL"} except that it only computes the
 probability mass function at a single outcome specified by \code{vec}.

\code{"NA"} is an approximation method using Normal approximation to 
compute the probability mass function of \code{vec} vector
specified by user.
}
\examples{
pp=matrix(c(.1, .1, .1, .7, .1, .3, .3, .3, .5, .2, .1, .2), nrow = 3, byrow = TRUE)


dpmd(pmat = pp)
dpmd(pmat = pp, method = "SIM-ALL", B = 10^3)
dpmd(pmat = pp, x = c(0,0,1,2), method = "NA" )
dpmd(pmat = pp, x = c(0,0,1,2), method = "SIM", B = 10^3)

}
