% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson-Multinomial_functions.R
\name{dpmd}
\alias{dpmd}
\title{Probability Mass of Poisson-Multinomial Distributions}
\usage{
dpmd(pp, method = "DFT-CF", vec = c(0, 0, 0, 0, 0), B = 100)
}
\arguments{
\item{pp}{a probability matrix}

\item{method}{method selected by user to compute the probability mass}

\item{vec}{result vec input by user}

\item{B}{simulation repeat time}
}
\value{
The probability mass of PMD
}
\description{
Probability Mass of Poisson-Multinomial Distributions
}
\examples{
aa=matrix(c(.1, .1, .1, .7, .1, .3, .3, .3, .5, .2, .1, .2, .5, .1, .1, .3), nrow=4, byrow=TRUE)
pp=aa[1:3,]
dpmd(pp)
}
