% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalstd.R
\name{getcluster}
\alias{getcluster}
\title{Get Pseudo-Spectrum as peaks cluster based on pmd analysis.}
\usage{
getcluster(list, corcutoff = NULL, accuracy = 4)
}
\arguments{
\item{list}{a list from getstd function}

\item{corcutoff}{cutoff of the correlation coefficient, default NULL}

\item{accuracy}{measured mass or mass to charge ratio in digits, default 4}
}
\value{
list with Pseudo-Spectrum index
}
\description{
Get Pseudo-Spectrum as peaks cluster based on pmd analysis.
}
\examples{
data(spmeinvivo)
re <- getpaired(spmeinvivo)
re <- getstd(re)
cluster <- getcluster(re)
}
\seealso{
\code{\link{getpaired}},\code{\link{getstd}},\code{\link{plotstd}}
}
