% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalstd.R
\name{getstd}
\alias{getstd}
\title{Identify standard ions through retention time clustering and PMD relationships}
\usage{
getstd(list, digits = 2, accuracy = 4)
}
\arguments{
\item{list}{A list object from getpaired() containing paired features}

\item{digits}{Rounding digits for mass differences}

\item{accuracy}{Mass accuracy for standard ion identification}
}
\value{
List with added standard ion indices and metadata
}
\description{
Identify standard ions through retention time clustering and PMD relationships
}
\examples{
data(spmeinvivo)
pmd <- getpaired(spmeinvivo)
std <- getstd(pmd)
}
\seealso{
\code{\link{getpaired}},\code{\link{getsda}},\code{\link{plotstd}}
}
