% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmdb}
\alias{hmdb}
\title{A dataframe containing HMDB top 10000 unique accurate mass pmd and related reactions}
\format{A dataframe with atoms numbers of C, H, O, N, P, S
\describe{
  \item{percentage}{accuracy of atom numbers prediction}
  \item{pmd}{pmd with two digits}
  }}
\usage{
data(hmdb)
}
\description{
A dataframe containing HMDB top 10000 unique accurate mass pmd and related reactions
}
\keyword{datasets}
