% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{l.vec.compute}
\alias{l.vec.compute}
\title{a function to compute l.vec}
\usage{
l.vec.compute(k, cn.vec, m)
}
\arguments{
\item{m, cn.vec, k}{categories and cn.vec}
}
\value{
l.vec a vector
}
\description{
a function to compute l.vec
}
