% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MaConDa}
\alias{MaConDa}
\title{mass spectrometry contaminants database for PMD check}
\format{
A data frame from \doi{doi:10.1093/bioinformatics/bts527} with 308 rows and 5 variables:
\describe{
  \item{id}{MaConDa ID}
  \item{name}{contaminants}
  \item{formula}{contaminants fomula}
  \item{exact_mass}{exact mass of contaminants}
  \item{type_of_contaminant}{type of contaminant}
  }
}
\usage{
data(MaConDa)
}
\description{
mass spectrometry contaminants database for PMD check
}
\keyword{datasets}
