% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pmatrix}
\alias{pmatrix}
\title{Title}
\usage{
pmatrix(n, m)
}
\arguments{
\item{n}{column dimension}

\item{m}{row dimension}
}
\value{
a randomly generated Poisson multinomial distribution probability matrix
}
\description{
Title
}
\examples{
pmatrix(2,2)

}
