% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson-Multinomial_functions.R
\name{ppmd}
\alias{ppmd}
\title{Cumulative Distribution Function of Poisson-Multinomial Distribution}
\usage{
ppmd(pmat, x, method = "DFT-CF", B = 1000)
}
\arguments{
\item{pmat}{A matrix of probabilities. Each row of pmat should add up 
to 1.}

\item{x}{Vector \eqn{x = (x_{1},x_{2},\ldots)} for computing 
\eqn{P(X_{1} \leq x_{1},X_{2} \leq x_{2},\ldots)}.}

\item{method}{Character string stands for the method selected by user to 
compute the probability mass. The method can only be one of 
the following three: 
\code{"DFT-CF"},
\code{"NA"},
\code{"SIM-ALL"}.}

\item{B}{Simulation repeating time. Will be ignored if users do not
choose \code{"SIM-ALL"} as method.}
}
\value{
The value of \eqn{P(X_{1} \leq x_{1},X_{2} \leq x_{2},\ldots)} of given 
\eqn{x = (x_{1},x_{2},\ldots)}.
}
\description{
This function computes cumulative distribution function of 
Poisson-Multinomial distributions that specified by input probability matrix 
via given method.
}
\details{
Three methods are same as listed in the details of \code{dpmd} but \code{"NA"}
stands for normal approximation. 

\code{ppmd} computes the cumulative distribution function by adding all probability 
mass points within hyper-dimensional space limited by \code{x}.
}
\examples{
pp=matrix(c(.1, .1, .1, .7, .1, .3, .3, .3, .5, .2, .1, .2), nrow=3, byrow=TRUE)

ppmd(pmat = pp, x = c(3,2,1,3))
ppmd(pmat = pp, x = c(3,2,1,3), method = "NA")
ppmd(pmat = pp, x = c(3,2,1,3), method = "SIM-ALL", B = 10^3)
}
