% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson-Multinomial_functions.R
\name{rpmd}
\alias{rpmd}
\title{generate random number from PMD}
\usage{
rpmd(pp)
}
\arguments{
\item{pp}{input matrix of probabilities}
}
\value{
the random number vector generated from PMD
}
\description{
generate random number from PMD
}
\examples{
aa=matrix(c(.1, .1, .1, .7, .1, .3, .3, .3, .5, .2, .1, .2, .5, .1, .1, .3), nrow=4, byrow=TRUE)
rpmd(pp=aa[1:3,])
}
