% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poisson-Multinomial_functions.R
\name{rpmd}
\alias{rpmd}
\title{Poisson-Multinomial Distribution Random Number Generator}
\usage{
rpmd(n, pmat)
}
\arguments{
\item{n}{Number of samples to be generated.}

\item{pmat}{A matrix of probabilities. Each row of pmat should add up 
to 1.}
}
\value{
A matrix of samples, each row stands for one sample.
}
\description{
Generating random samples of given a Poisson-Multinomial distribution.
}
\examples{
pp=matrix(c(.1, .1, .1, .7, .1, .3, .3, .3, .5, .2, .1, .2), nrow = 3, byrow = TRUE)


rpmd(n = 5, pmat = pp)

}
