% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sda}
\alias{sda}
\title{A dataset containing common Paired mass distances of substructure, ions replacements, and reaction}
\format{
A data frame with 94 rows and 4 variables:
\describe{
  \item{PMD}{Paired mass distances}
  \item{origin}{potential sources}
  \item{Ref.}{references}
  \item{mode}{b for biological reaction and e for environmental reaction}
  }
}
\usage{
data(sda)
}
\description{
A dataset containing common Paired mass distances of substructure, ions replacements, and reaction
}
\keyword{datasets}
